Add files via upload
An expansion of Rock, Paper, Scissors in the structure of Rock, Paper, Scissors, Lizard, Spock from the TV series The Big Bang Theory. Players can enter their choice in any variation of capital and lowercase, and the computer will randomly select it's choice. Options included to do one round or best of three, with score tracking and print messages of why you won/lost that round.